requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
djoser~=2.2.2
phonenumbers==8.13.26